backend/requirements.txt
flask==2.3.2
python-binance==1.0.19
pandas==2.0.1
numpy==1.24.3
ta-lib==0.4.20
python-telegram-bot==20.3
requests==2.31.0
python-dotenv==1.0.0
flask-cors==4.0.0
